Split file into multiple parsers
