+++ CommandLine/src/Worms.Armageddon.Resources/Replays/Text/IReplayLineParser.cs
@@ -1,0 +1,8 @@
+namespace Worms.Armageddon.Resources.Replays.Text
+{
+    internal interface IReplayLineParser
+    {
+        public bool CanParse(string line);
+        public void Parse(string line, ReplayResourceBuilder builder);
+    }
+}
