+++ CommandLine/src/Worms.Armageddon.Resources/Replays/Text/Parsers/WinnerParser.cs
@@ -1,0 +1,28 @@
+using System.Text.RegularExpressions;
+
+namespace Worms.Armageddon.Resources.Replays.Text.Parsers
+{
+    internal class WinnerParser : IReplayLineParser
+    {
+        private const string TeamName = @"(.+)";
+        private static readonly Regex WinnerDraw = new($"The round was drawn.");
+        private static readonly Regex Winner = new($"{TeamName} wins the (match!|round.)");
+
+        public bool CanParse(string line) => WinnerDraw.IsMatch(line) || Winner.IsMatch(line);
+
+        public void Parse(string line, ReplayResourceBuilder builder)
+        {
+            var winnerDrawMatch = WinnerDraw.Match(line);
+            var winnerMatch = Winner.Match(line);
+
+            if (winnerDrawMatch.Success)
+            {
+                builder.WithWinner("Draw");
+            }
+            else if (winnerMatch.Success)
+            {
+                builder.WithWinner(winnerMatch.Groups[1].Value);
+            }
+        }
+    }
+}
